refactor: split up large viz.py and add documentation
