+++ docs/api/viz.rst
@@ -1,0 +1,11 @@
+.. currentmodule:: tidy3d
+
+Visualization and Plotting
+==========================
+
+.. autosummary::
+    :toctree: _autosummary/
+    :template: module.rst
+
+    tidy3d.reset_previous_style
+    tidy3d.VisualizationSpec
