+++ tidy3d/components/viz/styles.py
@@ -1,41 +1,38 @@
-import matplotlib.pyplot as plt
-from matplotlib import cycler
-
-from ...log import log
-from .color_paletes import CATEGORICAL_PALETTES_HEX
-
-# Tidy3D default plotting style parameters
-_TIDY3D_STYLE_PARAMS = {
-    "axes.prop_cycle": cycler(color=CATEGORICAL_PALETTES_HEX[""]),
-    "axes.grid": True,
-    "grid.linestyle": ":",
-}
-
-# Store original parameters before applying Tidy3D style
-_ORIGINAL_PARAMS = {}
-
 try:
-    # Store only the parameters that Tidy3D style will modify
-    for key in _TIDY3D_STYLE_PARAMS:
-        if key in plt.rcParams:
-            _ORIGINAL_PARAMS[key] = plt.rcParams[key]
-    # Apply the Tidy3D style automatically on import
-    plt.rcParams.update(_TIDY3D_STYLE_PARAMS)
-except Exception as e:
-    log.error(f"Failed to apply Tidy3D plotting style on import. Error: {e}")
-    _ORIGINAL_PARAMS = {}  # Clear original params if application failed
+    from matplotlib.patches import ArrowStyle
 
+    arrow_style = ArrowStyle.Simple(head_length=11, head_width=9, tail_width=4)
+except ImportError:
+    arrow_style = None
 
-def reset_previous_style():
-    """
-    Resets matplotlib rcParams to the values they had before the Tidy3D
-    style was automatically applied on import.
-    """
-    if not _ORIGINAL_PARAMS:
-        log.warning("No previous Matplotlib style state found to reset to.")
-        return
+FLEXCOMPUTE_COLORS = {
+    "brand_green": "#00643C",
+    "brand_tan": "#B8A18B",
+    "brand_blue": "#6DB5DD",
+    "brand_purple": "#8851AD",
+    "brand_black": "#000000",
+    "brand_orange": "#FC7A4C",
+}
+ARROW_COLOR_SOURCE = FLEXCOMPUTE_COLORS["brand_green"]
+ARROW_COLOR_POLARIZATION = FLEXCOMPUTE_COLORS["brand_tan"]
+ARROW_COLOR_MONITOR = FLEXCOMPUTE_COLORS["brand_orange"]
+PLOT_BUFFER = 0.3
+ARROW_ALPHA = 0.8
+ARROW_LENGTH = 0.3
 
-    try:
-        plt.rcParams.update(_ORIGINAL_PARAMS)
-    except Exception as e:
-        log.error(f"Failed to reset previous Matplotlib style. Error: {e}")
+# stores color of simulation.structures for given index in simulation.medium_map
+MEDIUM_CMAP = [
+    "#689DBC",
+    "#D0698E",
+    "#5E6EAD",
+    "#C6224E",
+    "#BDB3E2",
+    "#9EC3E0",
+    "#616161",
+    "#877EBC",
+]
+
+# colormap for structure's permittivity in plot_eps
+STRUCTURE_EPS_CMAP = "gist_yarg"
+STRUCTURE_EPS_CMAP_R = "gist_yarg_r"
+STRUCTURE_HEAT_COND_CMAP = "gist_yarg"
